extract outbound resource tracking
