+++ crates/apps/reticulumd/src/bin/reticulumd/main.rs
@@ -8,6 +8,7 @@ mod bridge_helpers;
 mod inbound_worker;
 mod interface_hot_apply;
 mod interfaces;
+mod outbound_resources;
 mod receipt_events;
 mod receipt_worker;
 mod rpc_loop;
